@@ -1,4 +1,4 @@
-class CreateProjects < ActiveRecord::Migration
+class CreateProjects < ActiveRecord::Migration[6.0]
   def change
     create_table :projects do |t|
       t.string :name
