@@ -1,4 +1,4 @@
-class CreateEntries < ActiveRecord::Migration
+class CreateEntries < ActiveRecord::Migration[6.0]
   def change
     create_table :entries do |t|
       t.text :description
