@@ -1,4 +1,4 @@
-class CreateProjectLeaderboards < ActiveRecord::Migration
+class CreateProjectLeaderboards < ActiveRecord::Migration[6.0]
   def change
     create_table :project_leaderboards do |t|
       t.integer :project_id
