@@ -1,4 +1,4 @@
-class AddEmailToUser < ActiveRecord::Migration
+class AddEmailToUser < ActiveRecord::Migration[6.0]
   def change
     add_column :users, :email, :string
   end
