@@ -1,4 +1,4 @@
-class AddStateToUser < ActiveRecord::Migration
+class AddStateToUser < ActiveRecord::Migration[6.0]
   def change
     add_column :users, :state, :string
   end
